IAT-480: Refactor item service to use new directory structure.
