+++ src/app/core/core.module.ts
@@ -3,7 +3,7 @@ import {CommonModule} from "@angular/common";
 
 import {AlertService} from "./services/alert/alert.service";
 import {Logger} from "./services/logger/logger.service";
-import {ItemService} from "../item/services/item.service";
+import {ItemService} from "../item/services/item/item.service";
 import {ItemTypeService} from "../item/services/item-type/item-type.service";
 import {UserService} from "./services/user/user.service";
 import {HttpModule} from "@angular/http";
