+++ src/app/item/components/item-history/item-history.component.spec.ts
@@ -5,7 +5,7 @@ import {ItemHistoryComponent} from "./item-history.component";
 import {AppModule} from "../../../root/app.module";
 import {Logger} from "../../../core/services/logger/logger.service";
 import {AlertService} from "../../../core/services/alert/alert.service";
-import {ItemService} from "../../services/item.service";
+import {ItemService} from "../../services/item/item.service";
 
 describe('ItemHistoryComponent', () => {
   let component: ItemHistoryComponent;
