+++ src/app/item/services/item/item.service.spec.ts
@@ -3,7 +3,7 @@ import {HttpModule} from "@angular/http";
 import {ToastyModule} from "ng2-toasty";
 
 import {ItemService} from "./item.service";
-import {Logger} from "../../core/services/logger/logger.service";
+import {Logger} from "../../../core/services/logger/logger.service";
 
 describe('ItemService', () => {
   beforeEach(() => {
